chore: rename and delete card fikes
